// src/app/main/ecommerce/models/delivery/DeliveryBoyModel.ts
export interface DeliveryBoyModel {
	int_glcode: string;
	fk_vendor: string;
	var_name: string;
	var_email: string;
	var_mobile_no: string;
	var_password: string;
	var_address: string;
	var_latitude: string;
	var_longitude: string;
	chr_status: string;
	current_status: string;
	var_profile: string;
	nation_id: string;
	var_aadharcard: string;
	vehicle_image: string;
	var_pancard: string;
	var_device_token: string;
	chr_publish: string;
	chr_delete: string;
	dt_createddate: string;
	dt_modifydate: string;
	var_ipaddress: string;
	date: Date;
}